requirements.txt
# Enhanced CIFAR-10 CNN Project Dependencies - Academic Excellence Version (FIXED)
# Core ML Framework
tensorflow>=2.15.0
keras>=3.0.0

# Data Processing and Visualization - FIXED FOR COMPATIBILITY
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Computer Vision - FIXED VERSION FOR NUMPY COMPATIBILITY
opencv-python-headless>=4.8.0,<4.10.0
pillow>=10.0.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Web Interface
flask>=3.0.0

# Additional fixes for dependency conflicts
protobuf>=3.20.0,<5.0.0
werkzeug>=3.0.0

# Additional Enhancement Libraries
scipy>=1.11.0
joblib>=1.3.0
